requirements.txt
Pillow==8.3.2
pycodestyle==2.7.0
toml==0.10.2
